Ostoskori: Tuotteiden lisääminen ja poistaminen ostoskorista.

+++ src/main/java/fi/haagahelia/pizza/domain/Tuote.java
@@ -45,7 +45,7 @@ public class Tuote {
 		super();
 	}
 
-	public Tuote(int id, String name, int price) {
+	public Tuote(int id, String name, double price) {
 		this.id = id;
 		this.name = name;
 		this.price = price;
